+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Drive.java
@@ -60,8 +60,8 @@ public class Drive extends Control {
         }
 
         // Driving
-        double driveX = gamepad1.left_stick_x;
-        double driveY = -gamepad1.left_stick_y;
+        float driveX = gamepad1.left_stick_x;
+        float driveY = -gamepad1.left_stick_y;
         DriveMode driveMode = DriveMode.LOCAL;
 
         if(assistDrive) {
