+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tests/ConeStackTest.java
@@ -17,8 +17,8 @@ public class ConeStackTest extends AbstractAuto {
     public void loop() {
         super.loop();
 
-        double driveX = gamepad1.left_stick_x;
-        double driveY = -gamepad1.left_stick_y;
+        float driveX = gamepad1.left_stick_x;
+        float driveY = -gamepad1.left_stick_y;
         DriveMode driveMode = DriveMode.LOCAL;
 
         driveMode = DriveMode.GLOBAL;
